requirements.txt
mock==1.0.1
nose==1.3.4
PyYAML==3.11
requests==2.6.0
responses==0.3.0
six==1.9.0
